drm/i915: add minimal i915_gem_object_frontbuffer.h
commit c1464a89e1a41695d5bda47d5c27923b1802ce00 upstream.

Split out frontbuffer related declarations and static inlines from
gem/i915_gem_object.h into new gem/i915_gem_object_frontbuffer.h.

The main goal is to reduce header interdependencies. With
gem/i915_gem_object.h including display/intel_frontbuffer.h,
modification of the latter causes a whopping 300+ objects to be rebuilt,
while many of the source files actually needing it aren't explicitly
including it at all.

After the change, only 21 objects depend on display/intel_frontbuffer.h,
directly or indirectly.

deepin-Intel-SIG: commit c1464a89e1a4 drm/i915: add minimal i915_gem_object_frontbuffer.h.
MTL Graphics backporting - 2024Q2

Cc: Jouni Högander <jouni.hogander@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230830085127.2416842-1-jani.nikula@intel.com
[ Quanxian Wang: amend commit log ]
